backend/pyproject.toml
[project]
name = "ghost-job-detector-backend"
version = "0.1.0"
description = "FastAPI backend for ghost job detection with NLP analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",  
    "beautifulsoup4",
    "rapidfuzz",
    "supabase",
    "python-dotenv>=1.0.0",
    "ddgs"
]
